+++ src/fheroes2/dialog/dialog_selectfile.cpp
@@ -56,6 +56,7 @@
 #include "ui_keyboard.h"
 #include "ui_scrollbar.h"
 #include "ui_text.h"
+#include "ui_tool.h"
 #include "ui_window.h"
 #include "world.h"
