+++ src/fheroes2/gui/ui_keyboard.cpp
@@ -45,6 +45,7 @@
 #include "ui_button.h"
 #include "ui_language.h"
 #include "ui_text.h"
+#include "ui_tool.h"
 #include "ui_window.h"
 
 namespace
