+++ src/fheroes2/gui/ui_tool.cpp
@@ -24,11 +24,9 @@
 #include <array>
 #include <cassert>
 #include <cmath>
-#include <cstddef>
 #include <cstdlib>
 #include <cstring>
 #include <ctime>
-#include <string>
 #include <utility>
 
 #include "cursor.h"
@@ -654,6 +652,35 @@ namespace fheroes2
                 ++frameNumber;
             }
         }
+    }
+
+    size_t getTextInputCursorPosition( const std::string & text, const FontType & fontType, const size_t currentTextCursorPosition, const int32_t pointerCursorXOffset,
+                                       const int32_t textStartXOffset )
+    {
+        if ( text.empty() || pointerCursorXOffset <= textStartXOffset ) {
+            // The text is empty or mouse cursor position is to the left of input field.
+            return 0;
+        }
+
+        const int32_t maxOffset = pointerCursorXOffset - textStartXOffset;
+        const size_t textSize = text.size();
+        int32_t positionOffset = 0;
+        const CharHandler charHandler( fontType );
+
+        for ( size_t i = 0; i < textSize; ++i ) {
+            positionOffset += charHandler.getWidth( static_cast<uint8_t>( text[i] ) );
+
+            if ( positionOffset > maxOffset ) {
+                return i;
+            }
+
+            // If the mouse cursor is to the right of the current text cursor position we take its width into account.
+            if ( i == currentTextCursorPosition ) {
+                positionOffset += charHandler.getWidth( '_' );
+            }
+        }
+
+        return textSize;
     }
 
     void InvertedFadeWithPalette( Image & image, const Rect & roi, const Rect & excludedRoi, const uint8_t paletteId, const int32_t fadeTimeMs, const int32_t frameCount )
